+++ demo/admin/src/polyfills.ts
@@ -1,3 +1,0 @@
-// tslint:disable:no-submodule-imports
-import "react-app-polyfill/ie11";
-import "react-app-polyfill/stable";