+++ packages/admin/admin/src/theme/componentsTheme/MuiDialog.ts
@@ -1,4 +1,4 @@
-import { type Components } from "@mui/material/styles/components";
+import { type Components } from "@mui/material/styles";
 
 import { mergeOverrideStyles } from "../utils/mergeOverrideStyles";
 import { type GetMuiComponentTheme } from "./getComponentsTheme";